fixed bug #1504772 - Sources/Platform.h type definitions are not 64bit-safe
git-svn-id: https://svn.code.sf.net/p/nsis/code/NSIS@4712 212acab6-be3b-0410-9dea-997c60f758d6

+++ trunk/Source/Platform.h
@@ -16,7 +16,7 @@
 // basic types
 typedef unsigned char BYTE, *PBYTE, *LPBYTE;
 typedef unsigned short WORD, *LPWORD;
-typedef unsigned long DWORD, *LPDWORD;
+typedef unsigned int DWORD, *LPDWORD;
 typedef short SHORT;
 typedef unsigned short USHORT;
 typedef unsigned int UINT;
